Specific handling of OLD servers in ERROR state
VMs which end up in ERROR state are actually those which
can be purged very soon, at least in fully automated tenants.

(Unlike instances in powered down state, which may be part of
workflow).

So this adds new explicit option for age of servers in error.

Rebased and applied to master after re-facter
had to adjust unittests to keep test functionality